requirement.txt
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
google-api-core==2.25.1
google-auth==2.40.3
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-documentai==3.5.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
more-itertools==10.7.0
oauthlib==3.3.1
pdf2image==1.17.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
zipp==3.23.0
